Se definen permisos para Usuarios

+++ SIRMAT/Usuario/views.py
@@ -63,36 +63,43 @@ class UsuariosAPI(APIView):
         # Logica para peticiones tipo PUT
         # Es necesario proporcionar un parametro llamado 'id' con el valor del idcUsuario que se desea actualizar
         # ejmpl: usuarios/?id=1
+      
+        #Revisamos que la accion sea realizada por un super usuario o si el usuario se edita si mismo
+        if request.user.is_superuser or str(request.user.id) == request.query_params['id']:
+                  
+            if request.query_params:  # Revisamos si hay o no parametros dentro de la peticion HTTP
+                # Si los hay intentamos encontrar el elemento que coincida con el parametro 'id'
+                try:
+                    usuario = User.objects.get(id=request.query_params['id'])
+                # Si el try falla mandamos una respuesta con el error y un mensaje con detalles
+                except:
+                    return Response({
+                        'message': 'No hay parametro con nombre "id" o No se encontro ningun elemento que coincida con ese id'
+                    },  status=status.HTTP_404_NOT_FOUND)
 
-        if request.query_params:  # Revisamos si hay o no parametros dentro de la peticion HTTP
-            # Si los hay intentamos encontrar el elemento que coincida con el parametro 'id'
-            try:
-                usuario = User.objects.get(id=request.query_params['id'])
-            # Si el try falla mandamos una respuesta con el error y un mensaje con detalles
-            except:
+                # Si el try no falla entonces creamos el serializador utilizando el objeto guardado en 'usuario'
+                serializer = UsuarioSerializer(usuario, data=request.data)
+
+                if serializer.is_valid():  # Si la peticion es valida
+                    # Actualizamos el serializador en la base de datos y en label studio
+                    serializer.save()
+                    editar_usuario_ls(usuario.id, usuario.email,
+                                        request.data['password'])
+                    # Y respondemos con los datos del nuevo objeto creado
+                    return Response(serializer.data)
+                else:  # Si la peticion no es valida respondemos con un error y un mensaje con los detalles del error
+                    return Response(
+                        serializer.errors,
+                        status=status.HTTP_400_BAD_REQUEST
+                    )
+            else:  # El parametro es requerido por lo que si no se proporciona se respondera un error
                 return Response({
-                    'message': 'No hay parametro con nombre "id" o No se encontro ningun elemento que coincida con ese id'
-                },  status=status.HTTP_404_NOT_FOUND)
-
-            # Si el try no falla entonces creamos el serializador utilizando el objeto guardado en 'usuario'
-            serializer = UsuarioSerializer(usuario, data=request.data)
-
-            if serializer.is_valid():  # Si la peticion es valida
-                # Actualizamos el serializador en la base de datos y en label studio
-                serializer.save()
-                editar_usuario_ls(usuario.id, usuario.email,
-                                    request.data['password'])
-                # Y respondemos con los datos del nuevo objeto creado
-                return Response(serializer.data)
-            else:  # Si la peticion no es valida respondemos con un error y un mensaje con los detalles del error
-                return Response(
-                    serializer.errors,
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-        else:  # El parametro es requerido por lo que si no se proporciona se respondera un error
+                    'message': 'PUT debe proporcionar parametro "id"'
+                },  status=status.HTTP_400_BAD_REQUEST)
+        else:   
             return Response({
-                'message': 'PUT debe proporcionar parametro "id"'
-            },  status=status.HTTP_400_BAD_REQUEST)
+                "error" : "El usuario no tiene permisos para realizar esta accion"
+            },  status=status.HTTP_403_FORBIDDEN)
 
     # --------------------------------------------------------------------------------------------------------------
 
@@ -101,33 +108,40 @@ class UsuariosAPI(APIView):
         # Es necesario proporcionar un parametro llamado 'id' con el valor del idcUsuario que se desea eliminar
         # ejmpl: usuarios/?id=1
 
-        if request.query_params:  # Revisamos si hay o no parametros dentro de la peticion HTTP
+        #Revisamos que la accion sea realizada por un super usuario o si el usuario se edita si mismo
+        if request.user.is_superuser or str(request.user.id) == request.query_params['id']:
 
-            # Si los hay intentamos encontrar el elemento que coincida con el parametro 'id' y lo eliminamos
-            try:
-                usuario = User.objects.get(id=request.query_params['id'])
+            if request.query_params:  # Revisamos si hay o no parametros dentro de la peticion HTTP
 
-            # Si el try falla mandamos una respuesta con el error y un mensaje con detalles
-            except:
+                # Si los hay intentamos encontrar el elemento que coincida con el parametro 'id' y lo eliminamos
+                try:
+                    usuario = User.objects.get(id=request.query_params['id'])
 
+                # Si el try falla mandamos una respuesta con el error y un mensaje con detalles
+                except:
+
+                    return Response({
+                        'message': 'No hay parametro con nombre "id" o No se encontro ningun elemento que coincida con ese id'
+                    },  status=status.HTTP_404_NOT_FOUND)
+
+                # Si el try no falla entonces cambiamos el registro is_active de la BD
+
+                usuario.is_active = False #cambiamos is_active a False
+                usuario.save(update_fields = ['is_active']) #guardamos los cambios
+                eliminar_usuario_ls(usuario.id) #eliminamos usuario de LabelStudio
+                # Enviamos mensaje de éxito
                 return Response({
-                    'message': 'No hay parametro con nombre "id" o No se encontro ningun elemento que coincida con ese id'
-                },  status=status.HTTP_404_NOT_FOUND)
+                    'message': 'Usuario eliminado correctamente'
+                }, status=status.HTTP_200_OK)
 
-            # Si el try no falla entonces cambiamos el registro is_active de la BD
-
-            usuario.is_active = False #cambiamos is_active a False
-            usuario.save(update_fields = ['is_active']) #guardamos los cambios
-            eliminar_usuario_ls(usuario.id) #eliminamos usuario de LabelStudio
-            # Enviamos mensaje de éxito
+            else:  # El parametro es requerido por lo que si no se proporciona se respondera un error
+                return Response({
+                        'message' : 'DELETE debe proporcionar parametro "id"'
+                },  status = status.HTTP_400_BAD_REQUEST)
+        else:   
             return Response({
-                'message': 'Usuario eliminado correctamente'
-            }, status=status.HTTP_200_OK)
-
-        else:  # El parametro es requerido por lo que si no se proporciona se respondera un error
-            return Response({
-                    'message' : 'DELETE debe proporcionar parametro "id"'
-            },  status = status.HTTP_400_BAD_REQUEST)
+                "error" : "El usuario no tiene permisos para realizar esta accion"
+            },  status=status.HTTP_403_FORBIDDEN)
 
 
 class UsuariosSingUp(APIView):
